+++ bot/src/ChampionsOfKhazad.Bot.Lore/RelatedLoreService.cs
@@ -4,7 +4,7 @@ namespace ChampionsOfKhazad.Bot.Lore;
 
 internal class RelatedLoreService(IStoreLore loreStore, EmbeddingsService embeddingsService) : IGetRelatedLore
 {
-    public async Task<IReadOnlyList<Lore>> GetRelatedLoreAsync(string text, uint max = 10)
+    public async Task<IReadOnlyList<Lore>> GetRelatedLoreAsync(string text, ushort max = 10)
     {
         var embedding = await embeddingsService.CreateEmbeddingAsync(text);
